Move functions to a generic preprocessing module
